+++ src/Connector.DataLake.Common/Connector/SqlDataWriter/ISqlDataWriter.cs
@@ -1,12 +1,10 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
-
 using CluedIn.Core;
-
 using Microsoft.Data.SqlClient;
 
-namespace CluedIn.Connector.DataLake.Common.SqlDataWriter.Connector;
+namespace CluedIn.Connector.DataLake.Common.Connector.SqlDataWriter;
 
 internal interface ISqlDataWriter
 {
