+++ src/Connector.OneLake/OneLakeConnectorJobData.cs
@@ -3,60 +3,59 @@ using System.Collections.Generic;
 
 using CluedIn.Connector.DataLake.Common;
 
-namespace CluedIn.Connector.OneLake
+namespace CluedIn.Connector.OneLake;
+
+internal class OneLakeConnectorJobData : DataLakeJobData
 {
-    internal class OneLakeConnectorJobData : DataLakeJobData
+    public OneLakeConnectorJobData(
+        IDictionary<string, object> configurations,
+        string containerName = null)
+        : base(configurations, containerName)
     {
-        public OneLakeConnectorJobData(
-            IDictionary<string, object> configurations,
-            string containerName = null)
-            : base(configurations, containerName)
-        {
-        }
+    }
 
-        public string WorkspaceName => Configurations[OneLakeConstants.WorkspaceName] as string;
-        public string ItemName => Configurations[OneLakeConstants.ItemName] as string;
-        public string ItemType => Configurations[OneLakeConstants.ItemType] as string;
-        public string ItemFolder => Configurations[OneLakeConstants.ItemFolder] as string;
-        public string ClientId => Configurations[OneLakeConstants.ClientId] as string;
-        public string ClientSecret => Configurations[OneLakeConstants.ClientSecret] as string;
-        public string TenantId => Configurations[OneLakeConstants.TenantId] as string;
-        public override bool ShouldWriteGuidAsString => true;
+    public string WorkspaceName => Configurations[OneLakeConstants.WorkspaceName] as string;
+    public string ItemName => Configurations[OneLakeConstants.ItemName] as string;
+    public string ItemType => Configurations[OneLakeConstants.ItemType] as string;
+    public string ItemFolder => Configurations[OneLakeConstants.ItemFolder] as string;
+    public string ClientId => Configurations[OneLakeConstants.ClientId] as string;
+    public string ClientSecret => Configurations[OneLakeConstants.ClientSecret] as string;
+    public string TenantId => Configurations[OneLakeConstants.TenantId] as string;
+    public override bool ShouldWriteGuidAsString => true;
 
-        protected override void AddToHashCode(HashCode hash)
-        {
-            hash.Add(WorkspaceName);
-            hash.Add(ItemName);
-            hash.Add(ItemType);
-            hash.Add(ItemFolder);
-            hash.Add(ClientId);
-            hash.Add(ClientSecret);
-            hash.Add(TenantId);
+    protected override void AddToHashCode(HashCode hash)
+    {
+        hash.Add(WorkspaceName);
+        hash.Add(ItemName);
+        hash.Add(ItemType);
+        hash.Add(ItemFolder);
+        hash.Add(ClientId);
+        hash.Add(ClientSecret);
+        hash.Add(TenantId);
 
-            base.AddToHashCode(hash);
-        }
+        base.AddToHashCode(hash);
+    }
 
-        public override bool Equals(object obj)
-        {
-            return Equals(obj as OneLakeConnectorJobData);
-        }
+    public override bool Equals(object obj)
+    {
+        return Equals(obj as OneLakeConnectorJobData);
+    }
 
-        public bool Equals(OneLakeConnectorJobData other)
-        {
-            return other != null &&
-                WorkspaceName == other.WorkspaceName &&
-                ItemName == other.ItemName &&
-                ItemType == other.ItemType &&
-                ItemFolder == other.ItemFolder &&
-                ClientId == other.ClientId &&
-                ClientSecret == other.ClientSecret &&
-                TenantId == other.TenantId &&
-                base.Equals(other);
-        }
+    public bool Equals(OneLakeConnectorJobData other)
+    {
+        return other != null &&
+            WorkspaceName == other.WorkspaceName &&
+            ItemName == other.ItemName &&
+            ItemType == other.ItemType &&
+            ItemFolder == other.ItemFolder &&
+            ClientId == other.ClientId &&
+            ClientSecret == other.ClientSecret &&
+            TenantId == other.TenantId &&
+            base.Equals(other);
+    }
 
-        public override int GetHashCode()
-        {
-            return base.GetHashCode();
-        }
+    public override int GetHashCode()
+    {
+        return base.GetHashCode();
     }
 }
