+++ src/Connector.OneLake/OneLakeConnectorProvider.cs
@@ -3,24 +3,23 @@ using CluedIn.Core;
 using Microsoft.Extensions.Logging;
 using System.Collections.Generic;
 
-namespace CluedIn.Connector.OneLake
-{
-    public class OneLakeConnectorProvider : ConnectorProviderBase<OneLakeConnectorProvider>
-    {
-        public OneLakeConnectorProvider([NotNull] ApplicationContext appContext,
-            IOneLakeConstants configuration, ILogger<OneLakeConnectorProvider> logger)
-            : base(appContext, configuration, logger)
-        {
-        }
+namespace CluedIn.Connector.OneLake;
 
-        protected override IEnumerable<string> ProviderNameParts => new[]
-        {
-           OneLakeConstants.WorkspaceName,
-           OneLakeConstants.ItemFolder,
-           OneLakeConstants.ItemType,
-           OneLakeConstants.ItemName,
-           OneLakeConstants.ClientId,
-           OneLakeConstants.TenantId,
-        };
+public class OneLakeConnectorProvider : ConnectorProviderBase<OneLakeConnectorProvider>
+{
+    public OneLakeConnectorProvider([NotNull] ApplicationContext appContext,
+        IOneLakeConstants configuration, ILogger<OneLakeConnectorProvider> logger)
+        : base(appContext, configuration, logger)
+    {
     }
+
+    protected override IEnumerable<string> ProviderNameParts => new[]
+    {
+       OneLakeConstants.WorkspaceName,
+       OneLakeConstants.ItemFolder,
+       OneLakeConstants.ItemType,
+       OneLakeConstants.ItemName,
+       OneLakeConstants.ClientId,
+       OneLakeConstants.TenantId,
+    };
 }
